move code into recomemnders package
