+++ src/poprox_recommender/recommenders/load.py
@@ -1,8 +1,11 @@
+"""
+Functions and logic for loading recommender pipeline configurations.
+"""
+
 # pyright: basic
 import logging
-from typing import Any
 
-from lenskit.pipeline import Pipeline, PipelineBuilder, PipelineState
+from lenskit.pipeline import Pipeline, PipelineBuilder
 
 from poprox_concepts import CandidateSet, InterestProfile
 from poprox_recommender.components.diversifiers import (
@@ -32,33 +35,6 @@ class PipelineLoadError(Exception):
     Exception raised when a pipeline cannot be loaded or instantiated, to
     separate those errors from errors running the pipeline.
     """
-
-
-def select_articles(
-    candidate_articles: CandidateSet,
-    clicked_articles: CandidateSet,
-    interest_profile: InterestProfile,
-    pipeline_params: dict[str, Any] | None = None,
-) -> PipelineState:
-    """
-    Select articles with default recommender configuration.  It returns a
-    pipeline state whose ``default`` is the final list of recommendations.
-    """
-    available_pipelines = load_all_pipelines(device=default_device())
-    pipeline = available_pipelines["nrms"]
-
-    if pipeline_params and pipeline_params.get("pipeline"):
-        pipeline_name = pipeline_params["pipeline"]
-        pipeline = available_pipelines[pipeline_name]
-
-    recs = pipeline.node("recommender")
-    topk = pipeline.node("ranker", missing="none")
-    if topk is None:
-        wanted = (recs,)
-    else:
-        wanted = (topk, recs)
-
-    return pipeline.run_all(*wanted, candidate=candidate_articles, clicked=clicked_articles, profile=interest_profile)
 
 
 def load_all_pipelines(device: str | None = None, num_slots: int = 10) -> dict[str, Pipeline]:
